Renamed preprocess functions and fixed formatting
