+++ show_output.py
@@ -24,7 +24,7 @@ from PIL import Image
 from scipy import misc
 
 from fsc import get_fsc
-from preprocess_old import BLUE, WHITE, GRAY, BLACK, GREEN
+from preprocess_old import BLACK, BLUE, GRAY, GREEN, WHITE
 from train import build_net, load_inputs, load_masks
 
 # Time stamp of default image
@@ -43,137 +43,145 @@ WHITE_FOR_BLUE = [0, 170, 0]  # Medium green
 WHITE_FOR_GRAY = [0, 255, 0]  # Bright green
 RED = [255, 0, 0]
 
+
 def gather_images(times):
-    """"""
-    return [load_inputs(t) for t in times]
+	""""""
+	return [load_inputs(t) for t in times]
+
 
 def one_hot_to_mask(max_indices, output):
-    """Modifies (and returns) img to have sensible colors in place of
-    one-hot vectors."""
-    out = np.zeros([len(output), 480, 480 ,3])
-    out[(max_indices == 0)] = WHITE
-    out[(max_indices == 1)] = BLUE
-    out[(max_indices == 2)] = GRAY
-    out[(max_indices == 3)] = BLACK
-    out[(max_indices == 4)] = GREEN
-    return out
+	"""Modifies (and returns) img to have sensible colors in place of
+	one-hot vectors."""
+	out = np.zeros([len(output), 480, 480, 3])
+	out[(max_indices == 0)] = WHITE
+	out[(max_indices == 1)] = BLUE
+	out[(max_indices == 2)] = GRAY
+	out[(max_indices == 3)] = BLACK
+	out[(max_indices == 4)] = GREEN
+	return out
+
 
 def out_to_image(output):
-    """Modifies (and returns) the output of the network as a human-readable RGB
-    image."""
-    output = output.reshape([-1, 480, 480, 4])
-    # We use argmax instead of softmax so that we really will get one-hots
-    max_indexes = np.argmax(output, axis=3)
-    return one_hot_to_mask(max_indexes, output)
+	"""Modifies (and returns) the output of the network as a human-readable RGB
+	image."""
+	output = output.reshape([-1, 480, 480, 4])
+	# We use argmax instead of softmax so that we really will get one-hots
+	max_indexes = np.argmax(output, axis=3)
+	return one_hot_to_mask(max_indexes, output)
+
 
 def read_last_iteration_number(directory):
-    """Reads the output.txt file in directory. Returns the iteration number
-    on the last row."""
-    F = open(directory + 'output.txt', 'r')
-    file = F.readlines()
-    line = file[len(file) - 1]
-    return (line.split()[0])
+	"""Reads the output.txt file in directory. Returns the iteration number
+	on the last row."""
+	F = open(directory + 'output.txt', 'r')
+	file = F.readlines()
+	line = file[len(file) - 1]
+	return (line.split()[0])
+
 
 def read_parameters(directory):
-    """Reads the parameters.txt file in directory. Returns a dictionary
-    associating labels with keys."""
-    F = open(directory + 'parameters.txt', 'r')
-    file = F.readlines()
-    args = {}
-    for line in file:
-        key, value = line.split(':\t')
-        args[key] = value
-    return args
+	"""Reads the parameters.txt file in directory. Returns a dictionary
+	associating labels with keys."""
+	F = open(directory + 'parameters.txt', 'r')
+	file = F.readlines()
+	args = {}
+	for line in file:
+		key, value = line.split(':\t')
+		args[key] = value
+	return args
+
 
 def save_images(times, directory):
-    inputs = gather_images(times)
-    for i, input in enumerate(inputs):
-        input.save(directory + "input" + str(i) + ".jpg")
-
+	inputs = gather_images(times)
+	for i, input in enumerate(inputs):
+		input.save(directory + "input" + str(i) + ".jpg")
 
 
 def compare(output, target):
-    """Returns image of where output differs from target, color-coded by how
-    they agree or disagree. Destructively modifies target."""
-    target[np.logical_and((output == BLUE).all(axis=2),
-                          (target == GRAY).all(axis=2))] = RED
-    target[np.logical_and((output == BLUE).all(axis=2),
-                          (target == WHITE).all(axis=2))] = RED
-    target[np.logical_and((output == GRAY).all(axis=2),
-                          (target == BLUE).all(axis=2))] = RED
-    target[np.logical_and((output == GRAY).all(axis=2),
-                          (target == WHITE).all(axis=2))] = RED
-    target[np.logical_and((output == WHITE).all(axis=2),
-                          (target == BLUE).all(axis=2))] = RED
-    target[np.logical_and((output == WHITE).all(axis=2),
-                          (target == GRAY).all(axis=2))] = RED
-    target[np.logical_and((output == WHITE).all(axis=2),
-                          (target == BLACK).all(axis=2))] = RED
-    target[np.logical_and((output == BLACK).all(axis=2),
-                          (target == WHITE).all(axis=2))] = RED
-    target[np.logical_and((output == GRAY).all(axis=2),
-                          (target == BLACK).all(axis=2))] = RED
-    target[np.logical_and((output == BLACK).all(axis=2),
-                          (target == GRAY).all(axis=2))] = RED
-    target[np.logical_and((output == BLUE).all(axis=2),
-                          (target == BLACK).all(axis=2))] = RED
-    target[np.logical_and((output == BLACK).all(axis=2),
-                          (target == BLUE).all(axis=2))] = RED
-    # for i in range(target.shape[0]):
-    #     disp = Image.fromarray(targets[i].astype('uint8'))
-    #     if directory:
-    #         disp.save(directory + "compare" + str(i) + ".png")
-    #     else:
-    #         disp.show()
-    return Image.fromarray(target.astype('uint8'))
+	"""Returns image of where output differs from target, color-coded by how
+	they agree or disagree. Destructively modifies target."""
+	target[np.logical_and((output == BLUE).all(axis=2),
+			(target == GRAY).all(axis=2))] = RED
+	target[np.logical_and((output == BLUE).all(axis=2),
+			(target == WHITE).all(axis=2))] = RED
+	target[np.logical_and((output == GRAY).all(axis=2),
+			(target == BLUE).all(axis=2))] = RED
+	target[np.logical_and((output == GRAY).all(axis=2),
+			(target == WHITE).all(axis=2))] = RED
+	target[np.logical_and((output == WHITE).all(axis=2),
+			(target == BLUE).all(axis=2))] = RED
+	target[np.logical_and((output == WHITE).all(axis=2),
+			(target == GRAY).all(axis=2))] = RED
+	target[np.logical_and((output == WHITE).all(axis=2),
+			(target == BLACK).all(axis=2))] = RED
+	target[np.logical_and((output == BLACK).all(axis=2),
+			(target == WHITE).all(axis=2))] = RED
+	target[np.logical_and((output == GRAY).all(axis=2),
+			(target == BLACK).all(axis=2))] = RED
+	target[np.logical_and((output == BLACK).all(axis=2),
+			(target == GRAY).all(axis=2))] = RED
+	target[np.logical_and((output == BLUE).all(axis=2),
+			(target == BLACK).all(axis=2))] = RED
+	target[np.logical_and((output == BLACK).all(axis=2),
+			(target == BLUE).all(axis=2))] = RED
+	# for i in range(target.shape[0]):
+	#     disp = Image.fromarray(targets[i].astype('uint8'))
+	#     if directory:
+	#         disp.save(directory + "compare" + str(i) + ".png")
+	#     else:
+	#         disp.show()
+	return Image.fromarray(target.astype('uint8'))
+
 
 def show_output(accuracy, saver, x, y, y_, result_dir, num_iterations,
-             show_all, times, save=False):
-    """Loads the network and displays the output for the specified time."""
-    with tf.Session() as sess:
-        saver.restore(sess, result_dir + 'weights-' + str(num_iterations))
-        for i, t in enumerate(times):
-            input = load_inputs([t])
-            result = y.eval(feed_dict={x: input})
-            img = out_to_image(result)[0]
-            if show_all:
-                mask = np.array(misc.imread('data/simplemask/simplemask' + str(t) + '.png'))
-                input_image = Image.fromarray(input[0].astype('uint8'))
-                mask_image = Image.fromarray(mask.astype('uint8'))
-                comparison_image = compare(img, mask)
-            output_image = Image.fromarray(img.astype('uint8'))
-            if save:
-                output_image.save(result_dir + 'net-output' + str(i) + '.png')
-                if show_all:
-                    input_image.save(result_dir + 'input' + str(i) + '.jpg')
-                    mask_image.save(result_dir + 'mask' + str(i) + '.png')
-                    comparison_image.save(result_dir + 'compare' + str(i) + '.png')
-            else:
-                output_image.show()
-                if show_all:
-                        input_image.show()
-                        mask_image.show()
-                        comparison_image.show()
-            acc = accuracy.eval(feed_dict={x: input, y_: load_masks([t])})
-            print('Accuracy = ' + str(acc))
-            print("Network mask: ", get_fsc(times[0], img))
-            print("TSI mask: ", get_fsc(times[0], mask))
+		show_all, times, save=False):
+	"""Loads the network and displays the output for the specified time."""
+	with tf.Session() as sess:
+		saver.restore(sess, result_dir + 'weights-' + str(num_iterations))
+		for i, t in enumerate(times):
+			input = load_inputs([t])
+			result = y.eval(feed_dict={x: input})
+			img = out_to_image(result)[0]
+			if show_all:
+				mask = np.array(misc.imread('data/simplemask/simplemask' + str(t) + '.png'))
+				input_image = Image.fromarray(input[0].astype('uint8'))
+				mask_image = Image.fromarray(mask.astype('uint8'))
+				comparison_image = compare(img, mask)
+			output_image = Image.fromarray(img.astype('uint8'))
+			if save:
+				output_image.save(result_dir + 'net-output' + str(i) + '.png')
+				if show_all:
+					input_image.save(result_dir + 'input' + str(i) + '.jpg')
+					mask_image.save(result_dir + 'mask' + str(i) + '.png')
+					comparison_image.save(result_dir + 'compare' + str(i) + '.png')
+			else:
+				output_image.show()
+				if show_all:
+					input_image.show()
+					mask_image.show()
+					comparison_image.show()
+			acc = accuracy.eval(feed_dict={x: input, y_: load_masks([t])})
+			print('Accuracy = ' + str(acc))
+			print("Network mask: ", get_fsc(times[0], img))
+			print("TSI mask: ", get_fsc(times[0], mask))
+
 
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser()
-    parser.add_argument('directory')
-    parser.add_argument('--compare', help='If typed, images used to compare output will be displayed', action='store_true')
-    parser.add_argument('--time', help='Sets the time stamp to be loaded', type=int)
-    args = parser.parse_args()
-    time = TIME_STAMP
-    show_all = SHOW_ALL
-    if args.time:
-        time = args.time
-    if args.compare:
-        show_all = True
-    dir_name = "results/" + args.directory + "/"        
-    args = read_parameters(dir_name)
-    step_version = read_last_iteration_number(dir_name)
-    layer_info = args['Layer info'].split()
-    _, accuracy, saver, _, x, y, y_, _ = build_net(layer_info)
-    show_output(accuracy, saver, x, y, y_, dir_name, step_version, show_all, [time])
+	parser = argparse.ArgumentParser()
+	parser.add_argument('directory')
+	parser.add_argument('--compare', help='If typed, images used to compare output will be displayed',
+			action='store_true')
+	parser.add_argument('--time', help='Sets the time stamp to be loaded', type=int)
+	args = parser.parse_args()
+	time = TIME_STAMP
+	show_all = SHOW_ALL
+	if args.time:
+		time = args.time
+	if args.compare:
+		show_all = True
+	dir_name = "results/" + args.directory + "/"
+	args = read_parameters(dir_name)
+	step_version = read_last_iteration_number(dir_name)
+	layer_info = args['Layer info'].split()
+	_, accuracy, saver, _, x, y, y_, _ = build_net(layer_info)
+	show_output(accuracy, saver, x, y, y_, dir_name, step_version, show_all, [time])
